add navigation and restructure app
