+++ src/components/index.js
@@ -1,5 +1,7 @@
 import button from './Button';
 import signInForm from './SignInForm';
+import navigation from './Navigation';
 
 export const Button = button;
 export const SignInForm = signInForm;
+export const Navigation = navigation;
