+++ src/constants/routes.js
@@ -1,0 +1,6 @@
+export const routes = {
+    AUTH: '/auth',
+    HOME: '/dashboard',
+    ME: '/me',
+    CALENDAR: '/calendar',
+};
